refactor: extract shared module re-parsing into `module_reparse`
PR #173 (`import_granularity`) and this PR (`self_import`) independently
introduced the same late-pass machinery to reach out-of-line `mod foo;`
submodules: collect the source files that back a module in the HIR
module tree, build a throwaway `ParseSess` sharing the real `SourceMap`
with a `SilentEmitter`, and re-parse each file (which keeps `#[cfg(...)]`
gates intact). Now that both have landed on this branch via the master
merge, fold that duplicated code into a single crate-internal
`module_reparse` module exposing `for_each_module_file`, and have both
rules call it.

`self_import` also gains the explicit `parse_psess.edition` assignment
that #173 added, so edition-sensitive syntax re-parses exactly as the
crate compiles. No behavioral change otherwise; `just all` passes.
