+++ src/rules/import_granularity.rs
@@ -1,21 +1,12 @@
-use std::collections::HashSet;
-use std::sync::Arc;
-
 use clippy_utils::diagnostics::span_lint_hir_and_then;
 use clippy_utils::source::indent_of;
 use rustc_ast::{
-    AttrKind, Attribute, Crate, Item, ItemKind, MetaItemKind, ModKind, Visibility, VisibilityKind,
+    AttrKind, Attribute, Item, ItemKind, MetaItemKind, ModKind, Visibility, VisibilityKind,
 };
-use rustc_errors::emitter::SilentEmitter;
-use rustc_errors::{Applicability, DiagCtxt};
+use rustc_errors::Applicability;
 use rustc_lint::{LateContext, LateLintPass, LintContext, LintStore};
-use rustc_parse::lexer::StripTokens;
-use rustc_parse::new_parser_from_source_str;
-use rustc_session::parse::ParseSess;
 use rustc_session::{declare_tool_lint, impl_lint_pass};
-use rustc_span::def_id::LOCAL_CRATE;
-use rustc_span::source_map::SourceMap;
-use rustc_span::{BytePos, FileName, SourceFile, Span, sym};
+use rustc_span::{BytePos, Span, sym};
 
 mod check;
 mod config;
@@ -28,6 +19,7 @@ use model::{Leaf, StmtInfo, stmt_info};
 
 use crate::common::{DefaultState, resolved_state};
 use crate::enclosing_hir::find_enclosing_hir_ids;
+use crate::module_reparse::for_each_module_file;
 
 declare_tool_lint! {
     /// ### What it does
@@ -153,53 +145,13 @@ enum Violation {
 
 impl<'tcx> LateLintPass<'tcx> for ImportGranularity {
     fn check_crate(&mut self, lint_context: &LateContext<'tcx>) {
-        let tcx = lint_context.tcx;
-        let source_map = lint_context.sess().psess.clone_source_map();
-
-        // The files that define a module in this crate's module tree.
-        // Re-parsing is restricted to these so that source-map files
-        // which are *not* standalone modules — `include_str!` data and
-        // `include!` fragments — are not re-parsed and wrongly flagged.
-        let mut module_files: HashSet<FileName> = HashSet::new();
-        record_module_file(&source_map, &mut module_files, tcx.hir_root_module().spans);
-        for item_id in tcx.hir_free_items() {
-            if let rustc_hir::ItemKind::Mod(_, module) = &tcx.hir_item(item_id).kind {
-                record_module_file(&source_map, &mut module_files, module.spans);
-            }
-        }
-
-        // Snapshot the files before parsing: re-parsing takes a write
-        // lock on the shared source map, so it must not run while the
-        // `files()` read guard is held.
-        let module_source_files: Vec<Arc<SourceFile>> = {
-            let source_files = source_map.files();
-            source_files
-                .iter()
-                .filter(|source_file| source_file.cnum == LOCAL_CRATE)
-                .filter(|source_file| module_files.contains(&source_file.name))
-                .cloned()
-                .collect()
-        };
-
-        // A throwaway `ParseSess` sharing the real `SourceMap` (so spans,
-        // and our suggestions, point at the real files) but with a
-        // silenced `DiagCtxt`, so a file that does not parse cleanly
-        // standalone is skipped rather than surfacing parse errors.
-        let mut parse_psess = ParseSess::with_dcx(
-            DiagCtxt::new(Box::new(SilentEmitter)),
-            Arc::clone(&source_map),
-        );
-        // `with_dcx` already derives this from the root expansion (the
-        // crate's edition), but set it explicitly so edition-sensitive
-        // syntax re-parses exactly as the crate compiles.
-        parse_psess.edition = lint_context.sess().edition();
-
+        // Re-parse every module source file (reaching out-of-line
+        // submodules while keeping `#[cfg(...)]` gates intact) and check
+        // each file's items in turn. See [`crate::module_reparse`].
         let mut violations: Vec<Pending> = Vec::new();
-        for source_file in &module_source_files {
-            if let Some(krate) = parse_module_file(&parse_psess, source_file) {
-                self.check_items(lint_context, &krate.items, &mut violations);
-            }
-        }
+        for_each_module_file(lint_context, |krate| {
+            self.check_items(lint_context, &krate.items, &mut violations);
+        });
 
         if violations.is_empty() {
             return;
@@ -211,6 +163,7 @@ impl<'tcx> LateLintPass<'tcx> for ImportGranularity {
         // own span, not the merged replacement span: an out-of-line
         // `mod foo;` item's span lives in the parent file, so a merged
         // span there would fall back to the crate root.
+        let tcx = lint_context.tcx;
         let anchors: Vec<Span> = violations.iter().map(|pending| pending.anchor).collect();
         let hir_ids = find_enclosing_hir_ids(tcx, &anchors);
         for (pending, hir_id) in violations.into_iter().zip(hir_ids) {
@@ -243,62 +196,6 @@ impl<'tcx> LateLintPass<'tcx> for ImportGranularity {
                     }
                 },
             );
-        }
-    }
-}
-
-/// Record the on-disk source file that holds a module's body, keyed by
-/// name. A dummy span (no real body) contributes nothing. Only
-/// [`FileName::Real`] files count: a module synthesised by a proc macro
-/// has a `<proc-macro source>` file that must not be re-parsed and
-/// flagged as if the user wrote it.
-fn record_module_file(
-    source_map: &SourceMap,
-    module_files: &mut HashSet<FileName>,
-    spans: rustc_hir::ModSpans,
-) {
-    let inner_span = spans.inner_span;
-    if inner_span.is_dummy() {
-        return;
-    }
-    let name = &source_map.lookup_source_file(inner_span.lo()).name;
-    if matches!(name, FileName::Real(_)) {
-        module_files.insert(name.clone());
-    }
-}
-
-/// Re-parse a module's source file from its already-loaded text. Returns
-/// `None` (silently discarding buffered diagnostics — `parse_psess` is
-/// wired to a [`SilentEmitter`]) when the file does not parse as a
-/// standalone module. The shared source map already holds this file and
-/// deduplicates by name, so the parser reuses the loaded `SourceFile`
-/// (preserving the real spans) and the passed source text is ignored —
-/// hence the empty string, which avoids both a disk re-read and a clone
-/// of the whole file.
-fn parse_module_file(parse_psess: &ParseSess, source_file: &SourceFile) -> Option<Crate> {
-    // Load-bearing: a `SourceFile` without in-memory source makes the
-    // lexer ICE ("cannot lex `source_file` without source"). Local-crate
-    // `Real` files normally carry it, but bail rather than risk the ICE.
-    source_file.src.as_ref()?;
-    let mut parser = match new_parser_from_source_str(
-        parse_psess,
-        source_file.name.clone(),
-        String::new(),
-        StripTokens::ShebangAndFrontmatter,
-    ) {
-        Ok(parser) => parser,
-        Err(errors) => {
-            for error in errors {
-                error.cancel();
-            }
-            return None;
-        }
-    };
-    match parser.parse_crate_mod() {
-        Ok(krate) => Some(krate),
-        Err(error) => {
-            error.cancel();
-            None
         }
     }
 }
