+++ src/rules/self_import.rs
@@ -14,35 +14,26 @@
 //! - [`combined`] — the `combined` style's adjacency fold.
 //!
 //! The rule runs as a [`LateLintPass`] that **re-parses each of the
-//! crate's module source files** from a throwaway [`ParseSess`] sharing
-//! the real [`SourceMap`]. A pre-expansion pass would leave out-of-line
-//! `mod foo;` modules `ModKind::Unloaded` (their files are not read
-//! until macro expansion), so it would silently skip every separate-file
-//! submodule. Re-parsing reaches every submodule while keeping
-//! `#[cfg(...)]` gates intact (parsing does not strip cfg, unlike the
-//! post-expansion AST). This mirrors the sibling `import_granularity`
-//! rule.
-
-use std::collections::HashSet;
-use std::sync::Arc;
+//! crate's module source files** via [`crate::module_reparse`]. A
+//! pre-expansion pass would leave out-of-line `mod foo;` modules
+//! `ModKind::Unloaded` (their files are not read until macro expansion),
+//! so it would silently skip every separate-file submodule. Re-parsing
+//! reaches every submodule while keeping `#[cfg(...)]` gates intact
+//! (parsing does not strip cfg, unlike the post-expansion AST). The
+//! sibling `import_granularity` rule shares the same machinery.
 
 use clippy_utils::diagnostics::span_lint_hir_and_then;
 use rustc_ast::visit::{self, Visitor};
-use rustc_ast::{Block, Crate, Item, ItemKind, ModKind, Stmt, StmtKind};
-use rustc_errors::emitter::SilentEmitter;
-use rustc_errors::{Applicability, DiagCtxt};
+use rustc_ast::{Block, Item, ItemKind, ModKind, Stmt, StmtKind};
+use rustc_errors::Applicability;
 use rustc_hir::HirId;
-use rustc_lint::{LateContext, LateLintPass, LintContext, LintStore};
-use rustc_parse::lexer::StripTokens;
-use rustc_parse::new_parser_from_source_str;
-use rustc_session::parse::ParseSess;
+use rustc_lint::{LateContext, LateLintPass, LintStore};
 use rustc_session::{declare_tool_lint, impl_lint_pass};
-use rustc_span::def_id::LOCAL_CRATE;
-use rustc_span::source_map::SourceMap;
-use rustc_span::{FileName, SourceFile, Span};
+use rustc_span::Span;
 
 use crate::common::{DefaultState, resolved_state};
 use crate::enclosing_hir::find_enclosing_hir_ids;
+use crate::module_reparse::for_each_module_file;
 
 mod combined;
 mod forbid;
@@ -227,55 +218,20 @@ pub(super) enum Fix {
 
 impl<'tcx> LateLintPass<'tcx> for SelfImport {
     fn check_crate(&mut self, cx: &LateContext<'tcx>) {
-        let tcx = cx.tcx;
-        let source_map = cx.sess().psess.clone_source_map();
-
-        // The files that define a module in this crate's module tree.
-        // Re-parsing is restricted to these so source-map files that are
-        // *not* standalone modules — `include_str!` data, `include!`
-        // fragments — are not re-parsed and wrongly flagged.
-        let mut module_files: HashSet<FileName> = HashSet::new();
-        record_module_file(&source_map, &mut module_files, tcx.hir_root_module().spans);
-        for item_id in tcx.hir_free_items() {
-            if let rustc_hir::ItemKind::Mod(_, module) = &tcx.hir_item(item_id).kind {
-                record_module_file(&source_map, &mut module_files, module.spans);
-            }
-        }
-
-        // Snapshot the files before parsing: re-parsing takes a write
-        // lock on the shared source map, so it must not run while the
-        // `files()` read guard is held.
-        let module_source_files: Vec<Arc<SourceFile>> = {
-            let source_files = source_map.files();
-            source_files
-                .iter()
-                .filter(|source_file| source_file.cnum == LOCAL_CRATE)
-                .filter(|source_file| module_files.contains(&source_file.name))
-                .cloned()
-                .collect()
-        };
-
-        // A throwaway `ParseSess` sharing the real `SourceMap` (so spans,
-        // and our suggestions, point at the real files) but with a
-        // silenced `DiagCtxt`, so a file that does not parse cleanly
-        // standalone is skipped rather than surfacing parse errors.
-        let parse_psess = ParseSess::with_dcx(
-            DiagCtxt::new(Box::new(SilentEmitter)),
-            Arc::clone(&source_map),
-        );
-
+        // Re-parse every module source file (reaching out-of-line
+        // submodules while keeping `#[cfg(...)]` gates intact) and walk
+        // each file's scopes in turn. See [`crate::module_reparse`].
+        let style = self.style;
         let mut violations: Vec<Pending> = Vec::new();
-        for source_file in &module_source_files {
-            if let Some(krate) = parse_module_file(&parse_psess, source_file) {
-                let mut walker = SelfImportWalker {
-                    cx,
-                    style: self.style,
-                    violations: &mut violations,
-                };
-                walker.scan_items(krate.items.iter().map(|item| Some(&**item)));
-                visit::walk_crate(&mut walker, &krate);
-            }
-        }
+        for_each_module_file(cx, |krate| {
+            let mut walker = SelfImportWalker {
+                cx,
+                style,
+                violations: &mut violations,
+            };
+            walker.scan_items(krate.items.iter().map(|item| Some(&**item)));
+            visit::walk_crate(&mut walker, krate);
+        });
 
         if violations.is_empty() {
             return;
@@ -285,7 +241,7 @@ impl<'tcx> LateLintPass<'tcx> for SelfImport {
         // / per-item `#[allow]` resolves (emitting from `check_crate`
         // alone would sit at the crate root).
         let anchors: Vec<Span> = violations.iter().map(|pending| pending.anchor).collect();
-        let hir_ids = find_enclosing_hir_ids(tcx, &anchors);
+        let hir_ids = find_enclosing_hir_ids(cx.tcx, &anchors);
         for (pending, hir_id) in violations.into_iter().zip(hir_ids) {
             emit_pending(cx, hir_id, pending);
         }
@@ -323,62 +279,6 @@ fn emit_pending(cx: &LateContext<'_>, hir_id: HirId, pending: Pending) {
             }
         },
     );
-}
-
-/// Record the on-disk source file that holds a module's body, keyed by
-/// name. A dummy span (no real body) contributes nothing. Only
-/// [`FileName::Real`] files count: a module synthesised by a proc macro
-/// has a `<proc-macro source>` file that must not be re-parsed and
-/// flagged as if the user wrote it.
-fn record_module_file(
-    source_map: &SourceMap,
-    module_files: &mut HashSet<FileName>,
-    spans: rustc_hir::ModSpans,
-) {
-    let inner_span = spans.inner_span;
-    if inner_span.is_dummy() {
-        return;
-    }
-    let name = &source_map.lookup_source_file(inner_span.lo()).name;
-    if matches!(name, FileName::Real(_)) {
-        module_files.insert(name.clone());
-    }
-}
-
-/// Re-parse a module's source file from its already-loaded text. Returns
-/// `None` (silently discarding buffered diagnostics — `parse_psess` is
-/// wired to a [`SilentEmitter`]) when the file does not parse as a
-/// standalone module. The shared source map already holds this file and
-/// deduplicates by name, so the parser reuses the loaded `SourceFile`
-/// (preserving the real spans) and the passed source text is ignored —
-/// hence the empty string, which avoids both a disk re-read and a clone
-/// of the whole file.
-fn parse_module_file(parse_psess: &ParseSess, source_file: &SourceFile) -> Option<Crate> {
-    // Load-bearing: a `SourceFile` without in-memory source makes the
-    // lexer ICE ("cannot lex `source_file` without source"). Local-crate
-    // `Real` files normally carry it, but bail rather than risk the ICE.
-    source_file.src.as_ref()?;
-    let mut parser = match new_parser_from_source_str(
-        parse_psess,
-        source_file.name.clone(),
-        String::new(),
-        StripTokens::ShebangAndFrontmatter,
-    ) {
-        Ok(parser) => parser,
-        Err(errors) => {
-            for error in errors {
-                error.cancel();
-            }
-            return None;
-        }
-    };
-    match parser.parse_crate_mod() {
-        Ok(krate) => Some(krate),
-        Err(error) => {
-            error.cancel();
-            None
-        }
-    }
 }
 
 /// Walks one re-parsed module file, scanning each scope that holds a
